feat: PR comments, merge both service into one
